requirements.txt
attrs==25.3.0
beautifulsoup4==4.13.4
blinker==1.9.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
Flask==3.1.0
flask-cors==5.0.1
h11==0.14.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
numpy==2.2.5
outcome==1.3.0.post0
packaging==25.0
pandas==2.2.3
PySocks==1.7.1
python-dateutil==2.9.0.post0  # Keeping the latest version
python-dotenv==1.1.0
pytz==2025.2
requests==2.32.3  # Keep the latest version
schedule==1.2.2
selenium==4.31.0
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.7
trio==0.30.0
trio-websocket==0.12.2
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.4.0
webdriver-manager==4.0.2
websocket-client==1.8.0
Werkzeug==3.1.3
wsproto==1.2.0
Django==4.2.7
djangorestframework==3.14.0
django-filter==23.3
